requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
chromadb==1.0.13
fastapi==0.135.3
httpx==0.28.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-experimental==0.3.4
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.1
nemoguardrails==0.21.0
openai==1.86.0
pandas==2.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pytest==9.0.3
python-dotenv==1.1.0
starlette==1.0.0
streamlit==1.45.1
uvicorn==0.34.3